+++ packages/thumbprint-react/CHANGELOG.md
@@ -4,8 +4,12 @@
 
 -   [Patch] Simplify `<noscript>` output in `Image` component by removing `<picture>` and `srcSet` and using only `<img alt="..." src="..." />`.
 -   [Minor] Add the new `Popover` component.
--   [Patch] Refactor out shared ESC listener logic from popover and modal curtain.
--   [Patch] Refactor ModalCurtain to be a function component using hooks.
+-   [Patch] Create new hook for registering an ESC listener.
+-   [Patch] Create new hook for trapping focus.
+-   [Patch] Create new utility component for conditionally using portals.
+-   [Patch] Refactor ModalCurtain to be a function component and use hooks.
+-   [Patch] Refactor Tooltip to use the new hooks.
+-   [Patch] Refactor Popover to use the new hooks.
 
 ## 6.2.1 - 2019-06-24